chore: split custom wtforms objects by type
