+++ submission/utils/custom_forms.py
@@ -1,0 +1,99 @@
+""" Collection of custom form classes used throughout the submission system """
+
+from wtforms import Form, IntegerField, SelectField, validators
+from .custom_fields import TagListField
+from .custom_widgets import StructureInput
+
+
+class LocationForm(Form):
+    """Subform for location entry, use in combination with FormField"""
+
+    start = IntegerField(
+        "Start", validators=[validators.Optional(), validators.NumberRange(min=1)]
+    )
+    end = IntegerField(
+        "End", validators=[validators.Optional(), validators.NumberRange(min=2)]
+    )
+
+
+class EvidenceForm(Form):
+    method = SelectField(
+        "Method",
+        choices=[
+            "",
+            "Homology-based prediction",
+            "Correlation of genomic and metabolomic data",
+            "Gene expression correlated with compound production",
+            "Knock-out studies",
+            "Enzymatic assays",
+            "Heterologous expression",
+            "In vitro expression",
+        ],
+        validators=[validators.InputRequired()],
+    )
+    references = TagListField(
+        "Citation(s)",
+        [validators.InputRequired()],
+        description="Comma separated list of references. Accepted formats are: 'doi:10.1016/j.chembiol.2020.11.009', 'PMID:33321099'",
+    )
+
+
+class SubtrateEvidenceForm(Form):
+    name = SelectField(
+        "Method",
+        choices=[
+            "",
+            "Activity assay",
+            "ACVS assay",
+            "ATP-PPi exchange assay",
+            "Enzyme-coupled assay",
+            "Feeding study",
+            "Heterologous expression",
+            "Homology",
+            "HPLC",
+            "In-vitro experiments",
+            "Knock-out studies",
+            "Mass spectrometry",
+            "NMR",
+            "Radio labelling",
+            "Sequence-based prediction",
+            "Steady-state kinetics",
+            "Structure-based inference",
+            "X-ray crystallography",
+        ],
+    )
+    references = TagListField(
+        "Citation(s)", widget=StructureInput()
+    )  # TODO: standardize citations
+
+
+class StructureEvidenceForm(Form):
+    method = SelectField(
+        "Method",
+        choices=[
+            "",
+            "NMR",
+            "Mass spectrometry",
+            "MS/MS",
+            "X-ray crystallography",
+            "Chemical derivatisation",
+            "Total synthesis",
+        ],
+    )
+    references = TagListField(
+        "Citation(s)", widget=StructureInput()
+    )  # TODO: standardize citations
+
+
+class FunctionEvidenceForm(Form):
+    method = SelectField(
+        "Method",
+        choices=[
+            "",
+            "Other in vivo study",
+            "Heterologous expression",
+            "Knock-out",
+            "Activity assay",
+        ],
+    )
+    references = TagListField("Citation(s)", widget=StructureInput())
